removed process monitor singleton and added clear_server_singletons method to get integration tests working
