+++ tests/main/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+from pytest import register_assert_rewrite
+
+register_assert_rewrite("tests.helpers")
