+++ tests/test_main.py
@@ -2,31 +2,19 @@
 import base64
 import json
 import logging
-import multiprocessing
-import platform
-import sys
 import tempfile
 import threading
 import time
-from unittest.mock import ANY
 
 from freezegun import freeze_time
 from mantarray_desktop_app import CALIBRATION_NEEDED_STATE
 from mantarray_desktop_app import COMPILED_EXE_BUILD_TIMESTAMP
-from mantarray_desktop_app import CURRENT_SOFTWARE_VERSION
 from mantarray_desktop_app import get_api_endpoint
-from mantarray_desktop_app import get_mantarray_process_manager
-from mantarray_desktop_app import get_mantarray_processes_monitor
 from mantarray_desktop_app import get_server_port_number
-from mantarray_desktop_app import get_shared_values_between_server_and_monitor
-from mantarray_desktop_app import ImproperlyFormattedCustomerAccountUUIDError
-from mantarray_desktop_app import ImproperlyFormattedUserAccountUUIDError
 from mantarray_desktop_app import main
 from mantarray_desktop_app import MantarrayProcessesMonitor
-from mantarray_desktop_app import MultiprocessingNotSetToSpawnError
 from mantarray_desktop_app import prepare_to_shutdown
 from mantarray_desktop_app import process_monitor
-from mantarray_desktop_app import RecordingFolderDoesNotExistError
 from mantarray_desktop_app import server
 from mantarray_desktop_app import SERVER_READY_STATE
 from mantarray_desktop_app import SUBPROCESS_POLL_DELAY_SECONDS
@@ -56,43 +44,6 @@ __fixtures__ = [
     fixture_patched_xem_scripts_folder,
     fixture_test_process_manager_without_created_processes,
 ]
-
-
-@pytest.mark.timeout(15)
-def test_main_can_launch_server_with_no_args_from_entrypoint__default_exe_execution(
-    mocker, fully_running_app_from_main_entrypoint, patched_xem_scripts_folder
-):
-    spied_prepare_to_shutdown = mocker.spy(main, "prepare_to_shutdown")
-
-    _ = fully_running_app_from_main_entrypoint()
-    wait_for_subprocesses_to_start()
-
-    shutdown_response = requests.get(f"{get_api_endpoint()}shutdown")
-    assert shutdown_response.status_code == 200
-    confirm_port_available(
-        get_server_port_number(), timeout=5
-    )  # wait for shutdown to complete
-
-    spied_prepare_to_shutdown.assert_called_once_with()
-
-
-@pytest.mark.timeout(20)
-@pytest.mark.slow
-def test_main_entrypoint__correctly_assigns_shared_values_dictionary_to_process_monitor__and_sets_the_process_monitor_singleton(
-    fully_running_app_from_main_entrypoint, patched_xem_scripts_folder
-):
-    # Eli (3/11/20): there was a bug where we only passed an empty dict during the constructor of the ProcessMonitor in the main() function. So this test was created specifically to guard against that regression.
-    _ = fully_running_app_from_main_entrypoint([])
-    port = get_server_port_number()
-    confirm_port_in_use(port, timeout=5)
-    wait_for_subprocesses_to_start()
-    assert system_state_eventually_equals(CALIBRATION_NEEDED_STATE, 5) is True
-
-    the_process_monitor = get_mantarray_processes_monitor()
-    assert isinstance(the_process_monitor, MantarrayProcessesMonitor)
-
-    shared_values_dict = get_shared_values_between_server_and_monitor()
-    assert "in_simulation_mode" in shared_values_dict
 
 
 @pytest.fixture(
@@ -152,106 +103,6 @@ def test_main_can_launch_server_and_processes_and_initial_boot_up_of_ok_comm_pro
     )
 
 
-@pytest.mark.timeout(1)
-def test_main_argparse_debug_test_post_build(mocker):
-    # fails by hanging because Server would be started opened if not handled
-    mocker.patch.object(main, "configure_logging", autospec=True)
-    main.main(["--debug-test-post-build"])
-
-
-def test_main_configures_logging(mocker):
-    mocked_configure_logging = mocker.patch.object(
-        main, "configure_logging", autospec=True
-    )
-    main.main(["--debug-test-post-build"])
-    mocked_configure_logging.assert_called_once_with(
-        path_to_log_folder=None, log_file_prefix="mantarray_log", log_level=logging.INFO
-    )
-
-
-@pytest.mark.slow
-@pytest.mark.timeout(20)
-def test_main_configures_process_manager_logging_level__and_standard_logging_level__to_info_by_default(
-    mocker, fully_running_app_from_main_entrypoint, patched_xem_scripts_folder
-):
-    app_info = fully_running_app_from_main_entrypoint([])
-    mocked_configure_logging = app_info["mocked_configure_logging"]
-
-    mocked_configure_logging.assert_called_once_with(
-        path_to_log_folder=None, log_file_prefix=ANY, log_level=logging.INFO
-    )
-    process_manager = get_mantarray_process_manager()
-    process_manager_logging_level = process_manager.get_logging_level()
-    assert process_manager_logging_level == logging.INFO
-
-
-@pytest.mark.timeout(15)
-def test_main_configures_process_manager_logging_level__and_standard_logging_level__to_debug_when_command_line_arg_passed(
-    mocker, fully_running_app_from_main_entrypoint, patched_xem_scripts_folder
-):
-    app_info = fully_running_app_from_main_entrypoint(["--log-level-debug"])
-    mocked_configure_logging = app_info["mocked_configure_logging"]
-
-    mocked_configure_logging.assert_called_once_with(
-        path_to_log_folder=None, log_file_prefix=ANY, log_level=logging.DEBUG
-    )
-    process_manager = get_mantarray_process_manager()
-    process_manager_logging_level = process_manager.get_logging_level()
-    assert process_manager_logging_level == logging.DEBUG
-
-
-def test_main__raises_error_if_multiprocessing_start_method_not_spawn(mocker):
-    mocked_get_start_method = mocker.patch.object(
-        multiprocessing, "get_start_method", autospec=True, return_value="fork"
-    )
-    mocker.patch.object(main, "configure_logging", autospec=True)
-    with pytest.raises(MultiprocessingNotSetToSpawnError, match=r"'fork'"):
-        main.main(["--debug-test-post-build"])
-    mocked_get_start_method.assert_called_once_with(allow_none=True)
-
-
-def test_main__logs_system_info__and_software_version_at_very_start(
-    mocker, patched_shared_values_dict
-):
-    spied_info_logger = mocker.spy(main.logger, "info")
-    main_thread = threading.Thread(
-        target=main.main, args=[[]], name="thread_for_main_function_in_test"
-    )
-    main_thread.start()
-    confirm_port_in_use(
-        get_server_port_number(), timeout=3
-    )  # wait for server to boot up
-    requests.get(f"{get_api_endpoint()}shutdown")
-    main_thread.join()
-
-    assert CURRENT_SOFTWARE_VERSION in spied_info_logger.call_args_list[0][0][0]
-
-    uname = platform.uname()
-    uname_sys = getattr(uname, "system")
-    uname_release = getattr(uname, "release")
-    uname_version = getattr(uname, "version")
-    spied_info_logger.assert_any_call(f"System: {uname_sys}")
-    spied_info_logger.assert_any_call(f"Release: {uname_release}")
-    spied_info_logger.assert_any_call(f"Version: {uname_version}")
-    spied_info_logger.assert_any_call(f"Machine: {getattr(uname, 'machine')}")
-    spied_info_logger.assert_any_call(f"Processor: {getattr(uname, 'processor')}")
-    spied_info_logger.assert_any_call(f"Win 32 Ver: {platform.win32_ver()}")
-    spied_info_logger.assert_any_call(f"Platform: {platform.platform()}")
-    spied_info_logger.assert_any_call(f"Architecture: {platform.architecture()}")
-    spied_info_logger.assert_any_call(f"Interpreter is 64-bits: {sys.maxsize > 2**32}")
-    spied_info_logger.assert_any_call(
-        f"System Alias: {platform.system_alias(uname_sys, uname_release, uname_version)}"
-    )
-    spied_info_logger.assert_any_call(
-        f"Python Version: {platform.python_version_tuple()}"
-    )
-    spied_info_logger.assert_any_call(
-        f"Python Implementation: {platform.python_implementation()}"
-    )
-    spied_info_logger.assert_any_call(f"Python Build: {platform.python_build()}")
-    spied_info_logger.assert_any_call(f"Python Compiler: {platform.python_compiler()}")
-
-
 @pytest.mark.slow
 def test_main__calls_boot_up_function_upon_launch(
     patched_xem_scripts_folder,
@@ -287,60 +138,6 @@ def test_main__does_not_call_boot_up_function_upon_launch_if_command_line_arg_pa
 
     assert system_state_eventually_equals(SERVER_READY_STATE, 5) is True
     spied_boot_up.assert_not_called()
-
-
-def test_main__logs_command_line_arguments(mocker):
-    expected_command_line_args = ["--debug-test-post-build", "--log-level-debug"]
-    spied_info_logger = mocker.spy(main.logger, "info")
-    main_thread = threading.Thread(
-        target=main.main,
-        args=[expected_command_line_args],
-        name="thread_for_main_function_in_test",
-    )
-    main_thread.start()
-    main_thread.join()
-
-    spied_info_logger.assert_any_call(
-        "Command Line Args: {'debug_test_post_build': True, 'log_level_debug': True, 'skip_mantarray_boot_up': False, 'port_number': None, 'log_file_dir': None, 'initial_base64_settings': None}"
-    )
-
-
-def test_main__handles_base64_command_line_argument_with_padding_issue(mocker):
-    expected_command_line_args = [
-        "--debug-test-post-build",
-        "--initial-base64-settings=eyJyZWNvcmRpbmdfZGlyZWN0b3J5IjoiL2hvbWUvdWJ1bnR1Ly5jb25maWcvTWFudGFycmF5Q29udHJvbGxlci9yZWNvcmRpbmdzIn0",
-    ]
-    spied_info_logger = mocker.spy(main.logger, "info")
-    main.main(expected_command_line_args)
-
-    spied_info_logger.assert_any_call(
-        "Command Line Args: {'debug_test_post_build': True, 'log_level_debug': False, 'skip_mantarray_boot_up': False, 'port_number': None, 'log_file_dir': None, 'initial_base64_settings': 'eyJyZWNvcmRpbmdfZGlyZWN0b3J5IjoiL2hvbWUvdWJ1bnR1Ly5jb25maWcvTWFudGFycmF5Q29udHJvbGxlci9yZWNvcmRpbmdzIn0'}"
-    )
-
-
-@pytest.mark.slow
-def test_main__stores_and_logs_port_number_from_command_line_arguments(
-    mocker, patched_shared_values_dict
-):
-    spied_info_logger = mocker.spy(main.logger, "info")
-
-    expected_port_number = 1234
-    command_line_args = [f"--port_number={expected_port_number}"]
-    main_thread = threading.Thread(
-        target=main.main,
-        args=[command_line_args],
-        name="thread_for_main_function_in_test",
-    )
-    main_thread.start()
-    confirm_port_in_use(expected_port_number, timeout=3)
-    requests.get(f"{get_api_endpoint()}shutdown")
-    main_thread.join()
-
-    actual = get_server_port_number()
-    assert actual == expected_port_number
-    spied_info_logger.assert_any_call(
-        f"Using server port number: {expected_port_number}"
-    )
 
 
 @pytest.mark.slow
@@ -408,51 +205,6 @@ def test_main__stores_and_logs_user_settings_and_recordings_folder_from_command_
             patched_shared_values_dict["config_settings"]["Recording Directory"]
             == expected_recordings_dir
         )
-
-
-@pytest.mark.slow
-def test_main__raises_error_when_invalid_customer_account_uuid_is_passed_in_cmd_line_args(
-    mocker, patched_shared_values_dict
-):
-    invalid_uuid = "14b9294a-9efb-47dd"
-    test_dict = {
-        "customer_account_uuid": invalid_uuid,
-    }
-    json_str = json.dumps(test_dict)
-    b64_encoded = base64.urlsafe_b64encode(json_str.encode("utf-8")).decode("utf-8")
-    command_line_args = [f"--initial-base64-settings={b64_encoded}"]
-    with pytest.raises(ImproperlyFormattedCustomerAccountUUIDError, match=invalid_uuid):
-        main.main(command_line_args)
-
-
-@pytest.mark.slow
-def test_main__raises_error_when_invalid_user_account_uuid_is_passed_in_cmd_line_args(
-    mocker, patched_shared_values_dict
-):
-    invalid_uuid = "not a uuid"
-    test_dict = {
-        "user_account_uuid": invalid_uuid,
-    }
-    json_str = json.dumps(test_dict)
-    b64_encoded = base64.urlsafe_b64encode(json_str.encode("utf-8")).decode("utf-8")
-    command_line_args = [f"--initial-base64-settings={b64_encoded}"]
-    with pytest.raises(ImproperlyFormattedUserAccountUUIDError, match=invalid_uuid):
-        main.main(command_line_args)
-
-
-@pytest.mark.slow
-def test_main__raises_error_when_invalid_recording_folder_is_passed_in_cmd_line_args(
-    mocker, patched_shared_values_dict
-):
-    expected_recordings_dir = "fake directory"
-    test_dict = {
-        "recording_directory": expected_recordings_dir,
-    }
-    json_str = json.dumps(test_dict)
-    b64_encoded = base64.urlsafe_b64encode(json_str.encode("utf-8")).decode("utf-8")
-    command_line_args = [f"--initial-base64-settings={b64_encoded}"]
-    with pytest.raises(RecordingFolderDoesNotExistError, match=expected_recordings_dir):
-        main.main(command_line_args)
 
 
 @pytest.mark.slow
